Updated the piece set to have an array
Using the masks only is very quick for iterating through pieces, but if
we want to check what is on a square, it is quite slow.

Added an array in the Piece Set, that helps with indexing what piece is
present on a square from a square value.

Looks like there is quite a speed improvement.

Made a few various updates here and there too.