Refactor a bit engine search
Updated various places to use &[T] instead of &Vec<T>
Added a search Result that contains search results, sized to the
multi_pv setting
Updated the search to return SearchResults. Removed the move cache
ordering - Was expecting the bot to become stronger but it became weaker
it seems.

Updated all the tokio::spawn to avoid doing `let _ =`, as cargo clippy
disrecommends it.